asistente-rrhh/requirements.txt
python-dotenv
streamlit
langchain-core
langchain-community
langchain-openai
langchainhub
langchain-text-splitters
chromadb
unstructured[md]
lark
tiktoken